@@ -1,12 +1,11 @@
 {
   "name": "indexer",
-  "version": "0.7.0",
+  "version": "1.0.0",
   "scripts": {
     "clean:dist": "rimraf dist/*",
     "build": "npm run clean:dist && tsc && cp package.json dist",
     "prod": "node dist/main.js",
-    "start": "npm run build && concurrently \"tsc -w\" \"npm run watch\"",
-    "watch": "sleep 3 && nodemon --watch dist dist/main.js",
+    "start": "ts-node src/main.ts",
     "test:unit": "alsatian src/**/*.spec.ts",
     "lint": "tslint -c tslint.json 'src/**/*.ts'",
     "lint:fix": "tslint --fix -c tslint.json 'src/**/*.ts'",
@@ -14,42 +13,45 @@
     "check-coverage": "nyc npm run test"
   },
   "dependencies": {
-    "@sentry/node": "5.10.2",
+    "@irohalab/mira-shared": "^4.2.2",
+    "@sentry/node": "^10.5.0",
     "applicationinsights": "^1.8.7",
-    "axios": "^1.8.2",
-    "cheerio": "^1.0.0-rc.3",
-    "express": "^4.19.2",
-    "inversify": "^5.0.1",
-    "inversify-express-utils": "^6.3.2",
-    "mongodb": "^3.3.2",
+    "axios": "^1.11.0",
+    "cheerio": "^1.1.2",
+    "express": "^4.21.2",
+    "inversify": "^6.0.3",
+    "inversify-express-utils": "^6.5.0",
+    "mongodb": "^6.18.0",
+    "nanoid": "5.0.9",
     "parse-torrent": "^9.1.5",
-    "pg": "^7.7.1",
-    "puppeteer": "^1.11.0",
-    "reflect-metadata": "^0.1.12",
+    "pg": "^8.16.3",
+    "puppeteer": "^24.16.2",
+    "reflect-metadata": "^0.2.2",
     "saslprep": "^1.0.3",
-    "typescript": "^3.2.1",
-    "winston": "^3.3.3"
+    "winston": "^3.17.0"
   },
   "devDependencies": {
-    "@istanbuljs/nyc-config-typescript": "^0.1.3",
-    "@types/cheerio": "^0.22.16",
-    "@types/express": "^4.17.2",
-    "@types/mongodb": "^3.3.6",
-    "@types/node": "^10.12.12",
-    "@types/parse-torrent": "^5.8.4",
-    "@types/pg": "^7.4.11",
-    "@types/puppeteer": "^1.11.0",
-    "alsatian": "^3.0.1",
-    "concurrently": "^5.3.0",
-    "nodemon": "^2.0.4",
-    "nyc": "^14.1.1",
-    "rimraf": "^2.6.2",
-    "tslib": "^1.9.3",
-    "tslint": "^5.12.1"
+    "@istanbuljs/nyc-config-typescript": "^1.0.2",
+    "@types/amqplib": "^0.10.7",
+    "@types/express": "^4.17.23",
+    "@types/node": "^24.3.0",
+    "@types/parse-torrent": "^5.8.7",
+    "@types/pg": "^8.15.5",
+    "alsatian": "^3.2.1",
+    "concurrently": "^9.2.0",
+    "nyc": "^17.1.0",
+    "rimraf": "^6.0.1",
+    "ts-node": "^10.9.2",
+    "tslib": "^2.8.1",
+    "tslint": "^6.1.3",
+    "typescript": "^5.5.4"
   },
   "resolutions": {
     "minimist": "^0.2.1",
     "kind-of": "^6.0.3",
     "yargs-parser": "^13.1.2"
+  },
+  "engines": {
+    "node": ">20.0.0"
   }
 }
